Separate InstanceCreator from AbstractClassResolver
